Implemented base mathod for opg4, not tested properly yet. Refactored a bit

+++ src/MulShiftHash.cs
@@ -1,7 +1,7 @@
 public class MulShiftHash
 // This class implements a hash function that combines multiplication and bit shifting.
 {
-    private readonly Int64 a;
+    private readonly UInt64 a;
     private readonly UInt16 l; 
 
 
@@ -12,17 +12,17 @@ public class MulShiftHash
     {
         this.rnd = new Random();
         // Set the values of a and l
-        this.a = (rnd.NextInt64() << 1) - 1;
+        this.a = (UInt64) (rnd.NextInt64() << 1) - 1;
         this.l = l; 
     }
 
-    public UInt64 Hash(Int64 x)
+    public UInt64 Hash(UInt64 x)
     {
         // Perform a multiplication followed by a right shift
-        UInt64 leftside = (UInt64) (a * x);
+        UInt64 leftside = (a * x);
         int rightside = 64 - l;
         UInt64 rightshifted = leftside >> rightside; 
 
-        return (UInt64)rightshifted;
+        return rightshifted;
     }
 }
